+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added support for indexer parameter description. [#2978](https://github.com/microsoft/kiota/issues/2978)
+
 ### Changed
 
 ## [1.5.0] - 2023-08-04